[SYCL][E2E] Split vector scalar operator ordering test
This commit splits the vec_binary_scalar_order test to improve coverage
and reduce the compile-time overhead.
